+++ statsapi/poolsinfo.go
@@ -49,7 +49,7 @@ func poolsStatFetcher(sr *StatRequest, info *StatInfo) (results []StatResult, er
 		for _, id := range sr.EntityIDs {
 
 			// TODO - go somewhere and fetch values, capacity
-			values := []int{40, 27, 27, 34, 40, 90, 89, 50, 40, 30}
+			values := []float64{40, 27, 27, 34, 40, 90, 89, 50, 40, 30}
 			capacity := 100
 
 			if detailErr != nil {
